+++ src/utils.js
@@ -99,3 +99,27 @@ export function cloneArray(array) {
   if (!Array.isArray(array)) return [];
   return [].concat(array);
 }
+
+export function noop() {}
+
+// copied from dom-helpers
+let scrollbarSize;
+export function getScrollbarSize(recalculate) {
+  if ((!scrollbarSize && scrollbarSize !== 0) || recalculate) {
+    if (typeof window !== 'undefined' && window.document && window.document.createElement) {
+      let scrollDiv = document.createElement('div');
+
+      scrollDiv.style.position = 'absolute';
+      scrollDiv.style.top = '-9999px';
+      scrollDiv.style.width = '50px';
+      scrollDiv.style.height = '50px';
+      scrollDiv.style.overflow = 'scroll';
+
+      document.body.appendChild(scrollDiv);
+      scrollbarSize = scrollDiv.offsetWidth - scrollDiv.clientWidth;
+      document.body.removeChild(scrollDiv);
+    }
+  }
+
+  return scrollbarSize;
+}